[#57] 매도 매수 잔량 합 Double로 타입 변환

+++ BithumbMarket/BithumbMarket/View/Orderbook/OrderbookQuantitiesView.swift
@@ -55,7 +55,7 @@ final class OrderbookQuantitiesView: UIView {
         setConstraintLayout()
     }
     
-    func updateUI(sumOfAsks: Float, sumOfBids: Float) {
+    func updateUI(sumOfAsks: Double, sumOfBids: Double) {
         asksQuantitiesLabel.text = String(sumOfAsks).withDecimal(maximumDigit: 4)
         bidsQuantitiesLabel.text = String(sumOfBids).withDecimal(maximumDigit: 4)
     }
